repl: add :open to edit complex expressions in $EDITOR
Entering long multi-line expressions in the repl can be tedious.
There is currently no easy way to correct an error such as a missing
comma.
This commit adds :open, which will open a temporary file in $EDITOR
(much like `git commit`), allowing the user to perform arbitrary edits to the last executed
expression (empty if :open is the first command)
When $EDITOR exits, the file contents are ran in the repl.

Change-Id: I798df9176f116f748cc22d71629f31917f8d97a1